requirements.txt
streamlit
google-genai==0.16.0  